Munge ld.so's ELF headers so that sel_ldr will load it
We have to set e_type to ET_EXEC, because sel_ldr will refuse to load
ET_DYN objects.

However, the linker will refuse to treat ld.so as a library if it is
ET_EXEC, so we have to keep the original copy of ld.so around to link
against.  For a discussion, see:
http://code.google.com/p/nativeclient/issues/detail?id=156

We also have to remove the non-PT_LOAD ELF headers.  For a
discussion about this, see:
http://code.google.com/p/nativeclient/issues/detail?id=558

Review URL: http://codereview.chromium.org/2304004