Refactor makefile generation to make the build faster
glibc's Makerules generates a makefile called "sysd-rules" based on
the sysdirs search path.  Before this change, sysd-rules contains a
large number (over 2000) of wildcard rules such as:

$(objpfx)%.o: $(..)sysdeps/i386/elf/%.c $(before-compile)
        $(compile-command.c)

Unfortunately, GNU make's handling of wildcard rules is not efficient:
it takes quadratic time to add the rules to make's database, which
happens many times because glibc uses recursive make.

We change sysd-rules to use non-wildcard rules such as:

$(objpfx)start.os: $(..)sysdeps/i386/elf/start.S $(before-compile)
        $(compile-command.S)

Not only is this faster, it makes it easier to see which source files
the glibc build is using.  It will also make it easier to introduce a
mechanism that allows, for example, a subset of sysdeps/i386 to be
excluded from the search path, which is useful for nacl-glibc.

This reduces the time taken for the null build from 83s to 34s.

We also change the build's output to be less verbose.  Each
compile command, when printed, tends to take up half a windowful.
Instead, just print the target name.  This makes it easier to see
how the build is progressing.

Review URL: http://codereview.chromium.org/2319002