Remove excess page-alignment padding in executables
Remove the padding after the ELF headers that the linker previously
inserted to keep the code segment page-aligned in the file.  We do
this by adding SIZEOF_HEADERS so that the code segment's start is not
page-aligned.

To support this, change the dynamic loader to support non-page-aligned
code segment starts.  Previously it would have tried to load and
validate the ELF headers.

As a result, the hellow-dynamic executable is 2900 bytes stripped,
7350 bytes unstripped.

Previously it was 68180 bytes stripped, 72630 bytes unstripped.

Review URL: http://codereview.chromium.org/2823015