Make install tree relocatable: Remove absolute pathnames from libc.so
glibc's "libc.so" is a linker script (rather than the usual symbolic
link) that points to "libc.so.6".

libc.so normally contains the following, with absolute pathnames:

GROUP ( /lib/libc.so.6 /usr/lib/libc_nonshared.a AS_NEEDED ( /lib/ld-linux.so.2 ) )

That does not work if we want the development libraries to be stored
somewhere else for use at link time, in a relocatable location
relative to gcc or "-L". We change this to:

GROUP ( libc.so.6 libc_nonshared.a AS_NEEDED ( ld-linux.so.2 ) )

This is not ideal, because it will not work if there is an
incompatible "libc.so.6" in the current directory when linking.  This
is therefore a temporary change that we should revisit.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=597

Review URL: http://codereview.chromium.org/2854037