Remove most uses of the function attribute "regparm"
regparm(3) causes function arguments to be passed in registers.  We
cannot use this for functions that are called across dynamic objects
via the PLT because the dynamic linker's trampoline cannot preserve
all registers.  Whereas the normal i386 trampoline uses "ret $12" to
help preserve registers, this is not available under NaCl.

We introduce internal_asm_function for functions that are called from
assembly.  For now, this is simpler than changing the assembly code.

This is necessary for loading libpthread.so without it crashing, since
this calls ld.so's _dl_get_tls_static_info().

This is also necessary for libdl.so to work.

Review URL: http://codereview.chromium.org/2914005