Fix validation failures in most string functions
Switch to using the C implementations for most string functions,
instead of the assembly implementations.

Introduce a more convenient mechanism for doing this.  We list the
overrides in make_sysd_rules.py.  This replaces the override files in
sysdeps/nacl.

This is necessary for loading libpthread.so without crashing.  Add a
test for this.

Review URL: http://codereview.chromium.org/2957004