Fix brk() so that malloc() works for non-trivial allocations
Both the Linux and NaCl "brk" syscalls return the old break value if
the request failed, but the brk() wrapper is supposed to return an
error.  Fix the wrapper so that it does that, otherwise malloc.c does
not notice that the allocation failed and segfaults.

We expect brk() to fail early on, because, with our current
contiguous-code address space layout, we are mapping libraries' data
segments over the brk() area.  The malloc implementation is supposed
to fall back to using mmap() when brk() fails.

TEST=malloc(10*1024) is enough to reproduce this

Review URL: http://codereview.chromium.org/3135007