Allow building on case insensitive filesystems
Use the suffix ".ons" instead of ".oS", because the latter clashes
with ".os".  ".oS" was used for object files to be included in
libc_nonshared.a.  ".ons" is intended to suggest "nonshared".

This allows building on Mac OS X, which uses a case insensitive
filesystem by default.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=855
TEST="python build.py -b" on OS X with gawk and gsed installed from Mac Ports

Review URL: http://codereview.chromium.org/3299002