Fix SETUP_PIC_REG to pass the validator
Use nacljmp instead of ret.  Note that using NACLRET is not
appropriate here because it clobbers %ecx, which the caller might want
preserved.  Luckily the return address is exactly what this routine is
supposed to return.

We use the same instructions here for get_pc_thunk that nacl-gcc uses.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=876
TEST=run_float_math_test (with another change applied to fix RETs)

Review URL: http://codereview.chromium.org/3593001