Fix call to _dl_init so that environ is set correctly
An earlier change broke the call to _dl_init so that the argv/envp
arguments were the wrong way around.  This caused libc.so to set
environ to the wrong value in __libc_init_first(), although main()
receives argv via another route and so was still receiving the correct
values.

Swap the arguments and add comments to make the code clearer.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=862
TEST=run_fib_array_test (which relies on an env var being set)
  and the newly-added run_env_var_test

Review URL: http://codereview.chromium.org/3516007