Fix two validation failures in ld.so: disable check_consistency()
Remove a "ret" and a "ud2a" (both disallowed).

This is an odd macro.  It does a sanity check on %ebx, but there is no
hint at the call sites about why %ebx might be set incorrectly.  While
we could fix this to pass the validator, it is simpler to assume this
is old debugging code and remove it.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=876
TEST=small_tests --nacl_glibc

Review URL: http://codereview.chromium.org/3530010