Linker script: Place .eh_frame in the correct segment
This fixes "hello_world" to be small (~8k unstripped).  With the
.eh_frame section in the wrong segment, the linker introduces ~64k of
padding.  (For some reason, the earlier executable size test in
nacl/test.py was not introducing .eh_frame into the link.)

We leave the linker script for libraries for now, because this
requires further fixes, but we add some comments.

BUG=none
TEST=run_hello_world_size_test (to be added to the Scons tests)

Review URL: http://codereview.chromium.org/3560018