Ensure futex_emulation.c is not linked into libc.so
This object's thread list should be instantiated only once, in ld.so.
Listing it in "routines" causes this object to be linked into libc.so.
"dl-routines" is the correct place to put it.

Before this change, any direct use of contended locking in libc.so
will result in a mutex_lock(0) syscall which causes sel_ldr to print
warnings because this is not a valid lock descriptor.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=649
TEST=Found by running run_thread_test, though other problems remain.
  "nm out/install/glibc/nacl/lib/libc.so.6 | grep futex"
  should show the __nacl_futex_* symbols as undefined.

Review URL: http://codereview.chromium.org/3791008