Linker script: Fix for executables with large BSSes
Align the BSS to a 64k page boundary in order to work around the issue
NaCl's mmap() has with zero filling.

I will fix elf_nacl.xs in a later change: this is more complicated,
since it is also used for ld.so, which is loaded by sel_ldr.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1068
TEST=run_app_lib_test now passes (since this uses a large BSS);
  run_hello_world_size_test demonstrates the need to move the
    "DATA_SEGMENT_END" declaration.

Review URL: http://codereview.chromium.org/3882003