pthread_create: Ensure that the new thread's stack is properly aligned
Adding the attribute "force_align_arg_pointer" forces gcc to align the
stack, but without us having to write arch-specific code to do it
ourselves.

With this in place, we do not rely on sel_ldr forcing stack alignment
in the thread_create syscall.

In Linux glibc, this alignment was done in clone.S.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1043
TEST=run_stack_alignment_test with the alignment in
  sel_ldr's NaClCreateAdditionalThread() disabled

Review URL: http://codereview.chromium.org/3935002