Fix futexes: Ensure they get initialised only once
_dl_init() was not the right place to initialise futexes, because it
is called when new libraries are loaded, such as via dlopen().

Do the initialisation in dl_main() instead.  The choice of position in
this function is fairly arbitrary.

Add an assertion to catch this, just in case.

As a tidy up, remove the INIT_LIST_HEAD() call.  This was redundant
because the LIST_HEAD macro already does this initialisation
statically.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1075
TEST=a subset of run_thread_test

Review URL: http://codereview.chromium.org/4027003