Remove excess padding in x86-64 dynamically-linked executables
Adding "+ SIZEOF_HEADERS" allows the code segment to share a page with
the ELF headers, so that padding is not required between the two.

This brings the linker script into line with the x86-32 version.

This reduces hello_world.nexe from >2MB (excessive because of issue
1073) to 10916 bytes unstripped.

BUG=none
TEST=run_hello_world_size_test, although this will need to be updated
  because it asserts a size limit of 10000 bytes, and the x86-64
  executable is still bigger than the x86-32 equivalent, possibly
  because we're using ELFCLASS64.

Review URL: http://codereview.chromium.org/5762003