Make static linking work on x86-64
Add a linker script for static linking on x86-64.  This is the same as
elf_nacl.x.static but with different architecture names.

Fix up sysdeps/x86_64/elf/start.S to ignore %rdx and not treat it as
the address of an atexit() handler function, because sel_ldr does not
zero %rdx on process startup (although it could).  This brings this
file into line with sysdeps/i386/elf/start.S, which has a similar
modification.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=940
TEST=run_hello_world_test nacl_static_link=1 (plus scons changes)

Review URL: http://codereview.chromium.org/6099005