Update for changed NaCl syscall ABI with 64-bit off_t, ino_t and time_t
We change time_t as presented by nacl-glibc's ABI to be 64-bit.  This
is the easiest way to keep NACL_PRIdNACL_TIME matching up and avoids
having to convert struct timeval and struct timespec in the syscall
wrappers.

However, off_t and ino_t as presented by nacl-glibc's ABI remain the
same for now.

mktime.c: Remove a compile-time assertion that does not seem to make
sense because it fails when we increase the width of the output while
keeping the input widths the same.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1253
TEST=small_tests --nacl_glibc

Review URL: http://codereview.chromium.org/6272001