Add the ability to receive the argv/env arrays over IPC on startup
This will allow command line arguments and environment variables to be
passed to a process under the NaCl browser plugin without having to
extend the plugin.

This is complicated by the ELF ABI for argv/env, through which ld.so
is invoked and it, in turn, invokes the executable.  This ABI requires
argv/env to be passed on the stack and requires assembly code.  In a
later iteration we could change this ABI and simplify this code.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1270
TEST=to be posted separately

Review URL: http://codereview.chromium.org/6183001