Extend open() to work as an RPC call
Add the option for open() to work as an RPC call, over IMC sockets.
This is an alternative to using NaCl's "open" syscall which is usually
only available in a debug mode.

This will allow the dynamic linker to work inside the NaCl browser
plugin.

We need imc_sendmsg() and imc_recvmsg() to be linked into libpthread
now because libpthread links its own copy of open().

Move the NACL_PLUGIN_* FD numbers to a separate file, nacl_rpc.h.
Rename NACL_PLUGIN_ASYNC_RECEIVE_FD to NACL_PLUGIN_ASYNC_TO_CHILD_FD
to make it clearer which side is receiving.  Use the correct FD number
this time (previously it did not match the plugin).

BUG=http://code.google.com/p/nativeclient/issues/detail?id=607
TEST=to be posted separately

Review URL: http://codereview.chromium.org/2832052