Fix a validation failure in libpthread.so
Remove pt-vfork.  Before the change, it is supplied from
nptl/sysdeps/unix/sysv/linux/*/pt-vfork.S, which does not pass the
validator.

There is still a stub vfork() function in libc.so, so this does not
break programs that want to link against the stub.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=876
TEST=check output of ncval_annotate.py on libpthread.so.0

Review URL: http://codereview.chromium.org/6708034