Truncate x86-64 rtld's l_addr to 32 bits.
The vanilla ld.so code computes its own load address from the PC.
On x86-64 the whole PC value includes the %r15 base, which should
not be considered part of the address in NaCl.

Before this change, the l_addr of the dynamic linker itself is a large
value that does not fit into 32 bits.  This leads to "causes overflow in
R_X86_64_32 relocation" warnings when the dynamic linker does the
relocation of its own data segment, which contains 32-bit pointer fields.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1411
TEST=rtld warning messages disappear in hand runs

R=khim@google.com,pasko@google.com,eaeltsin@google.com

Review URL: http://codereview.chromium.org/6883188