Adjust for new NaCl startup ABI
This replaces the startup code for applications (crt1.o) and the dynamic
linker startup and glue code with versions adapted to the new NaCl startup
ABI.  This uses the normal C function-call ABI to start the e_entry address
as a function of one argument, that being an information block holding the
arguments, environment, etc.

This is to enable http://codereview.chromium.org/7276050/, which see.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1131
TEST=manually tested, bots can't test because of staging issues

R=pasko@google.com,eaeltsin@google.com,mseaborn@google.com

Review URL: http://codereview.chromium.org/7282019