Fix up glibc for drift in syscall and IRT interfaces
The glibc copy of the IRT interfaces is out of date.  Fix it up.  The glibc
copy of the pre-IRT syscall interfaces was also out of date (passing extra
arguments since removed), and with http://codereview.chromium.org/7593001/
that mismatch comes to matter.  Fix that up too.

BUG= cleanup
TEST= hand built glibc and ran nacl tests against it

R=khim@google.com,pasko@google.com

Review URL: http://codereview.chromium.org/7569023