Quick and dirty fix to green the bots.
Replaces movdqa wath movups. Perhaps we can keep it as permanent solution:
1. Athlon XP (without SSE2!) is still around (under name Geode NX). One example:
  http://www.netbookchoice.com/2010/06/10/asus-eee-pc-1201k-packs-an-amd-geode-processor/
2. Upstream comment in sysdeps/x86_64/dl-trampoline.S looks ominous enough:
  /* Even though the stack is correctly aligned to allow using movaps
     we use movups.  Some callers might provide an incorrectly aligned
     stack and we do not want to have it blow up here.  */
     movups %xmm0, 16(%rcx)
     movups %xmm1, 32(%rcx)

This comment is removed (and code rewamped significantly) in upstream trunk,
but we'll need to redo porting work anyway: code is significantly different
there (because of the AVX support and other reasons).

Commited TBR to green the bots.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2208
TEST= look on buildbots

Review URL: http://codereview.chromium.org/7835015