Use irt's open_resource only for files in /lib
When I've added "lib" prefix I was under impression that
open_resource should be used to open all kind of resources.

Such as:
  /lib/libc.so.VERSION - it's a libc binary itself
  /lib/gconv/gconv-modules - list of supported gconv modules
  /share/locale/locale.alias - locale aliases

But this plan was vetoed on Chrome side. Now resource_open
is supposed to be used exclusively for DSO. It makes no sense
to use directories in such a case.

Now we use different scheme:
  /lib/libpthread.so.VERSION is loaded using IRT's open_resource
  /usr/lib/libpthread.so.VERSION is loaded using regular open(2)

R=pasko@google.com

Review URL: http://codereview.chromium.org/7937005