Fix __jmp_buf size for x86-64
Because of the type used vs the ILP32 configuration, this type is too
small for the assembly code that uses it.  It's hard to notice in
straightforward tests because the public jmp_buf type effectively has
unused padding (the space for the signal state).  But the internal
type is used in the implementation of pthread_cleanup_push, such that
using that interface will clobber nearby stack.  Note that the C++
implementation and the C implementation used under -fexceptions don't
use this.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2490
TEST= toolchain trybot

R=elijahtaylor@google.com,pasko@google.com,khim@google.com,eaeltsin@google.com

Review URL: http://codereview.chromium.org/8937018