Remove alignment of code and data segments for shared libraries.
Unlike ld.so code and data segments in shared libraries do not need
to be aligned to page boundary.  So I remove the alignment from
linker scripts but return it back in elf/Makefile for ld.so.

Also fix ld.so to allocate ELF header in the heap if code segment
is not aligned to page boundary.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=1794
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3266
TEST= toolchain glibc try bots.

Review URL: https://codereview.chromium.org/11958039